backend/requirements.txt
fastapi
uvicorn
openai
qdrant-client
python-dotenv
psycopg2-binary
